Refactored initialization of debug names: - Added "debugName" field to all RenderSystemChild descriptors. - Replaced "name" field with "debugName" field in ShaderDescriptor. - Renamed SetName function to SetDebugName. - Added AnsiString class to C# wrapper to simplify conversion between C# strings and ANSI strings.
Misc:
- Fixed ShadowMapping example to only use 1 MIP level for shadow map.
- Fixed warning in C99 Texturing example.

+++ examples/C99/Texturing/Texturing.c
@@ -233,7 +233,7 @@ int main(int argc, char* argv[])
                 llglSetResource(2, LLGL_GET_AS(LLGLResource, samplers[0]));
 
                 // Draw cube mesh with index and vertex buffers
-                llglDrawIndexed(indexCount, 0);
+                llglDrawIndexed((uint32_t)indexCount, 0);
             }
             llglEndRenderPass();
         }
